% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fretboardr-package.R
\docType{package}
\name{fretboardr-package}
\alias{fretboardr}
\alias{fretboardr-package}
\title{fretboardr: An Amazing Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: firstname lastname \email{your@email.com}

}
\keyword{internal}
