% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fretboard.R
\name{thumb}
\alias{thumb}
\alias{thmb}
\alias{index}
\alias{indx}
\alias{middle}
\alias{mddl}
\alias{ring}
\alias{pinky}
\alias{pnky}
\alias{barre}
\alias{barr}
\alias{mute}
\title{Add points to a fretboard}
\usage{
thumb(
  str,
  fr,
  label = "T",
  shape = "diamond",
  colour = "black",
  fill = "black",
  size = 5
)

thmb(
  str,
  fr,
  label = "T",
  shape = "diamond",
  colour = "black",
  fill = "black",
  size = 5
)

index(
  str,
  fr,
  label = "I",
  shape = "circle",
  colour = "black",
  fill = "black",
  size = 5
)

indx(
  str,
  fr,
  label = "I",
  shape = "circle",
  colour = "black",
  fill = "black",
  size = 5
)

middle(
  str,
  fr,
  label = "M",
  shape = "circle",
  colour = "black",
  fill = "black",
  size = 5
)

mddl(
  str,
  fr,
  label = "M",
  shape = "circle",
  colour = "black",
  fill = "black",
  size = 5
)

ring(
  str,
  fr,
  label = "R",
  shape = "circle",
  colour = "black",
  fill = "black",
  size = 5
)

pinky(
  str,
  fr,
  label = "P",
  shape = "circle",
  colour = "black",
  fill = "black",
  size = 5
)

pnky(
  str,
  fr,
  label = "P",
  shape = "circle",
  colour = "black",
  fill = "black",
  size = 5
)

barre(
  str,
  fr,
  label = "",
  shape = "circle",
  colour = "black",
  fill = "black",
  size = 5
)

barr(
  str,
  fr,
  label = "",
  shape = "circle",
  colour = "black",
  fill = "black",
  size = 5
)

mute(str, fr = -1, shape = "cross", colour = "black", fill = "black", size = 5)
}
\arguments{
\item{str}{The string(s) to put the point on}

\item{fr}{The fret to put the point on. Use \code{-1} to place a point behind the
nut - this is common notation for muting.}

\item{label}{The label for the point}

\item{shape, colour, fill, size}{Used to change the appearance of the point -
passed to \code{ggplot2::geom_point()}}
}
\value{
A named list
}
\description{
Use \code{thmb()}, \code{indx()}, \code{mddl()}, \code{ring()} and \code{pnky()} to add points for the
different digits. Use \code{mute()} and \code{barr()} for special markers.
}
\details{
Aliases \code{thumb()}, \code{index()}, \code{middle()}, \code{pinky()} and \code{barre()} are
provided, but the four-character functions are nice for alignment.
}
