% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{diagram}
\alias{diagram}
\title{Make a fretboard diagram}
\usage{
diagram(
  title = NULL,
  ...,
  frets = NULL,
  strings = c("E1", "A", "D", "G", "B", "E2"),
  flip = c("none", "x", "y", "both"),
  rotate = FALSE,
  auto_open = TRUE,
  label_strings = FALSE,
  debug = FALSE
)
}
\arguments{
\item{title}{The title for the diagram}

\item{...}{Points to place on the fretboard - constructed using calls to
\code{fngr()}, \code{thmb()}, \code{barr()} and \code{mute()}}

\item{frets}{The number of frets to show}

\item{strings}{The number of strings to show}

\item{flip}{Mirroring of the output - either \code{"none"}, \code{"x"}, \code{"y"} or
\code{"both"}}

\item{rotate}{If \code{TRUE} the fretboard will be shown in a horizontal layout}

\item{auto_open}{If \code{TRUE}, strings which don't have any markers specified
will be marked as 'open'.}

\item{debug}{If \code{TRUE} some diagnostic stuff will be shown.}
}
\value{
A ggplot
}
\description{
Make a fretboard diagram
}
\examples{
diagram(
  "F# (Henrix Style)",
  thmb(1, 2),
  ring(2, 4),
  pnky(3, 4),
  mddl(4, 3),
  barr(5:6, 2, "I"),
  rotate = TRUE
)
}
