% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fretboard.R
\name{fretboard}
\alias{fretboard}
\title{Plot a fretboard}
\usage{
fretboard(
  title = NULL,
  frets = 5,
  strings = factor(c("E1", "A", "D", "G", "B", "E2"), c("E1", "A", "D", "G", "B", "E2")),
  label_strings = TRUE,
  flip = c("none", "x", "y", "both"),
  rotate = FALSE
)
}
\arguments{
\item{title}{A title for the chord chart}

\item{frets}{The number of frets to include}

\item{strings}{The number of strings to include}
}
\value{
A ggplot
}
\description{
This provides the base for each new chord chart
}
