% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fretboard.R
\name{fngr}
\alias{fngr}
\alias{finger}
\alias{thmb}
\alias{thumb}
\alias{barr}
\alias{barre}
\alias{mute}
\title{Add points to a fretboard}
\usage{
fngr(
  str,
  fr,
  label = "",
  shape = "circle",
  colour = "black",
  fill = "black",
  size = 5
)

finger(
  str,
  fr,
  label = "",
  shape = "circle",
  colour = "black",
  fill = "black",
  size = 5
)

thmb(
  str,
  fr,
  label = "T",
  shape = "diamond",
  colour = "black",
  fill = "black",
  size = 5
)

thumb(
  str,
  fr,
  label = "T",
  shape = "diamond",
  colour = "black",
  fill = "black",
  size = 5
)

barr(
  str,
  fr,
  label = "",
  shape = "circle",
  colour = "black",
  fill = "black",
  size = 5
)

barre(
  str,
  fr,
  label = "",
  shape = "circle",
  colour = "black",
  fill = "black",
  size = 5
)

mute(str, fr = -1, shape = "cross", colour = "black", fill = "black", size = 5)
}
\arguments{
\item{str}{The string(s) to put the point on}

\item{fr}{The fret to put the point on. Use \code{-1} to place a point behind the
nut - this is common notation for muting.}

\item{label}{The label for the point}

\item{shape, colour, fill, size}{Used to change the appearance of the point -
passed to \code{ggplot2::geom_point()}}
}
\value{
A named list
}
\description{
Aliases \code{fngr()}, \code{thmb()} and \code{barr()} are provided because code looks nicer
when things align.
}
